Moved animations from Player.cs to AnimationManager.cs. Seems like it works, but not too many tests were made.
